requirements.txt
Django==1.5
Pillow==2.2.1
South==0.7.6
django-allauth==0.8.3
django-bootstrap-form==3.0
django-bootstrap3-datetimepicker==1.0.3
django-crispy-forms==1.4.0
django-debug-toolbar==0.9.4
django-extensions==1.0.2
django-localflavor==1.0
django-registration==1.0
django-tastypie==0.10.0
django-xadmin==0.4.0
httplib2==0.7.7
oauth2==1.5.211
python-dateutil==2.1
python-mimeparse==0.1.4
python-openid==2.2.5
requests==1.0.4
six==1.4.1
wsgiref==0.1.2
yolk==0.4.3